Modified Test Query and spaces
